==> RNAseq_Analysis/Differential_Expression/DEseq2/4___Per_Treatment_Alone/Step_1__WaldTest__DEseq_Main_Using_Gene_Counts_For_Initial_Structured_Data_Files.ipynb <==
# Steps 1 Generating Files From Wald Test For 2 Days Post Fertilization (DPF) Samples Only

## Site(s) Used:

* For the [demo version](http://bioconductor.org/packages/devel/bioc/vignettes/DESeq2/inst/doc/DESeq2.html)

* The [installation of the DESeq2]( https://bioconductor.org/packages/release/bioc/html/DESeq2.html )


## Defining Output Paths

# Define the parent directory to output the Wald Test

parent_directory <- "/path/to/parent/folder/of/your/choice"


## Loading Counts Matrix


First the counts matrix needs to be loaded. Make sure it is the one without any summary statistics.

Make sure to use the full path below as an input for `counts_matrix_file`.

# Define the path to your counts matrix file
counts_matrix_file <- "/path/to/my/counts/matrix/file/"

# Read the counts matrix from the TSV file
counts_matrix <- read.table(counts_matrix_file, header = TRUE, row.names = 1, sep = "\t")

head(counts_matrix)

print(colnames(counts_matrix))

# Specify the columns you want to isolate by name
columns_to_isolate <- c("Choose",
"TheColumnsYouWantToIsolated")

# Isolate the specified columns and keep the row names
isolated_columns <- counts_matrix[, c(columns_to_isolate), drop = FALSE]

head(isolated_columns)



sample_metadata <- data.frame(
  Sample = c("C.01__Control",
             "C.02__Control",
             "C.03__Control",
             "E.01__Experimental",
             "E.02__Experimental",
             "E.03__Experimental"),
    
  Treatment = c("Untreated", "Untreated", "Untreated",
                "Knockdown", "Knockdown", "Knockdown")
)

# View the table

sample_metadata


Below is the tentative design formula I am going to use.

colnames(sample_metadata)

## DEseq2 Data Set

When it comes to the design forumla for the `DESeqDataSetFromMatrix` object ensure that you put the formula directly into design. Do not put it into a variable that will be fed in design. It will most likely not work.

This is the design with only treatment as a covariate. Timepoint is not a covariate here.

library(DESeq2)

dds <- DESeqDataSetFromMatrix(countData = isolated_columns,
                              colData = sample_metadata,
                              design = ~ Treatment  )


## Manually Identify: Factor Levels

You may have to idiftify the reference factor as according to DEseq2:

> By default, R will choose a reference level for factors based on alphabetical order. Then, if you never tell the DESeq2 functions which level you want to compare against (e.g. which level represents the control group), the comparisons will be based on the alphabetical order of the levels. There are two solutions: you can either explicitly tell results which comparison to make using the contrast argument (this will be shown later), or you can explicitly set the factors levels. In order to see the change of reference levels reflected in the results names, you need to either run DESeq or nbinomWaldTest/nbinomLRT after the re-leveling operation. Setting the factor levels can be done in two ways, either using factor

Therefore below I am setting Untreated as the reference.  This makes it seem the knockdown as the experimental group.

dds$Treatment <- factor(dds$Treatment, levels = c("Untreated", "Knockdown"))

dds$Treatment <- relevel(dds$Treatment, ref = "Untreated")

## Pre-filtering Low Counts

keep <- rowSums(counts(dds)) >= 10
dds <- dds[keep,]

## Differential Expression Analysis Main

dds <- DESeq(dds)
res <- results(dds)
res

## Manually Specify What You Want To Compare

Here I would like to compare the treatments, the untreated vs the treated.

### Wald Test (Default)

res <- results(dds, contrast=c("Treatment","Knockdown", "Untreated"))

The preference is that you order by the smallest p value.

resOrdered <- res[order(res$pvalue),]

df_resOrdered <- as.data.frame(resOrdered)
head(as.data.frame(resOrdered))

Now it is time to output the results from the Wald test. The file name will have the time in Hours, Minutes, Seconds and whether it is AM or PM for version control purposes.

# Create a timestamp with date and time
timestamp <- format(Sys.time(), format = "%Y_%m_%d_%I%M%S%p")

# Create the Wald test subdirectory under the parent directory
wald_test_subdirectory <- file.path(parent_directory, "1___Wald_Test")
dir.create(wald_test_subdirectory, showWarnings = FALSE)

# Define the full path for the Wald test file with date and time
Wald_test_file_path <- file.path(wald_test_subdirectory, paste0("Wald_Test_Contast_Treatment_Knockdown_Control_", timestamp, ".tsv"))

# Write the data frame to a TSV file
write.table(df_resOrdered, file = Wald_test_file_path, sep = "\t", row.names = TRUE)

### Examining Data:

summary(res)

To look at how many p values were less than 0.05 do the following command:

sum(res$padj < 0.05, na.rm=TRUE)

1598 values are there with a FDR less than 0.05 that is nice.

res05 <- results(dds, alpha=0.05)
summary(res05)

sum(res05$padj < 0.05, na.rm=TRUE)

resultsNames(dds)

## Log fold change shrinkage for visualization and ranking

resLFC <- lfcShrink(dds, coef="Treatment_Knockdown_vs_Untreated", type="apeglm")
resLFC

## Independent hypothesis weighting

library("IHW")
resIHW <- results(dds, filterFun=ihw)
summary(resIHW)
sum(resIHW$padj < 0.05, na.rm=TRUE)
metadata(resIHW)$ihwResult

## Data Exploration:

plotMA(res, ylim=c(-2,2))

plotMA(resLFC, ylim=c(-2,2))

## Session info

Currently I running this locally on the work mac.

sessionInfo()

print("Done")
